Speed up AINode IT and split CPU-only tests off GPU runner
Two-tier isolation for org.apache.iotdb.ainode.it so that only the
GPU-bound tests need a self-hosted GPU runner:

- New AINodeIT category for tests that exercise only metadata/lifecycle
  paths (SHOW/DROP builtin model, REMOVE AINODE). These tests run on a
  plain ubuntu-latest runner via the new AINodeIT Maven profile and
  cluster-it-ainode-cpu.yml workflow.
- AIClusterIT profile keeps tests that drive CALL INFERENCE, FORECAST,
  or LOAD MODEL TO DEVICES (CUDA), still running on the GPU runner.
- AINodeBasicIT collects the 4 metadata tests previously mixed into
  AINodeSharedClusterIT; AINodeClusterConfigIT is re-tagged to AINodeIT.
- AINodeWrapper now tolerates a missing /data/ainode/models cache, so
  CPU runners can boot AINode without the multi-GB weight bundle. When
  the cache is present, weights are symlinked instead of copied per fork
  to remove a large per-test-class IO cost.

Additional speedups in the GPU pipeline:

- AINodeConcurrentForecastIT loop count 100 -> 10 (still 100 reqs per
  model for a concurrency smoke check; nightly can dial up).
- AINodeTestUtils.prepareDataInTree/Table/Table2 and
  AINodeConcurrentForecastIT.prepareDataForTableModel switched from
  per-row execute() to addBatch()/executeBatch() in chunks of 500.
