+++ integration-test/pom.xml
@@ -694,6 +694,20 @@
             </properties>
         </profile>
         <profile>
+            <id>AINodeIT</id>
+            <activation>
+                <activeByDefault>false</activeByDefault>
+            </activation>
+            <properties>
+                <integrationTest.excludedGroups>org.apache.iotdb.itbase.category.ManualIT</integrationTest.excludedGroups>
+                <integrationTest.includedGroups>org.apache.iotdb.itbase.category.AINodeIT</integrationTest.includedGroups>
+                <integrationTest.launchNodeInSameJVM>false</integrationTest.launchNodeInSameJVM>
+                <integrationTest.randomSelectWriteNode>false</integrationTest.randomSelectWriteNode>
+                <integrationTest.readAndVerifyWithMultiNode>false</integrationTest.readAndVerifyWithMultiNode>
+                <integrationTest.testEnv>AI</integrationTest.testEnv>
+            </properties>
+        </profile>
+        <profile>
             <id>DailyIT</id>
             <activation>
                 <activeByDefault>false</activeByDefault>
